+++ packages/fetch/src/index.ts
@@ -1,4 +1,4 @@
-export * from './a/MockFetch';
+export * from './mock/MockFetch';
 export * from './Fetch';
 export * from './FetchError';
 export * from './IFetch';
