+++ packages/redis/src/mock/MockRedisList.ts
